Extract Partials generation (stages 6-8) into a new PartialsManager class
TestBuildExecutor owned every stage from preprocessing through
execution; the three Partials stages (preprocessing partial headers/
sources, then extracting and generating implementation/interface
content) are self-contained beyond the TestPipelineHelpers mixin
already shared across this directory's pipeline classes, and are 446
lines on their own -- pulled into their own class, wired into
TestPipelineManager alongside the other three pipeline-stage classes,
mirroring the TestSourceFileDirectiveResolver precedent. No behavior
change: the moved methods are unchanged, TestPipelineManager's stage
sequence, conditions, and stop-point behavior are unchanged, and every
existing unit example for the moved stages passed unmodified once
retargeted at the new class's own (smaller) constructor.
